--- src/controler/addProduto.js
@@ -1,3 +1,0 @@
-import { kknex } from "../db/database.js";
-
- //vai adicionar os produtos
