+++ src/controler/controlerDistribuidor.js
@@ -1,28 +1,13 @@
-import { kknex } from "../db/database.js";
+import { kknex } from "../config/database.js";
+import { insertUser } from "./controlerUser.js";
 
-
-//------------faz o cadastro do id geral--------------------------------------------
-
-async function insertUser(email, senha) {
-  try {
-    const [userId] = await kknex('Usuario').insert({
-      email,
-      senha,
-      tipo: 1
-    });
-    return userId;
-  } catch (error) {
-    console.error('Erro ao inserir usuário:', error);
-    throw error;
-  }
-}
 //------------cadastra os dados exclusivos--------------------------------------------
 
 async function insertDistributor(username, endereco, usuarioId) {
   try {
     const [distributorId] = await kknex('Distribuidores').insert({
-      username,
-      endereco,
+      username:username,
+      endereco:endereco,
       usuario_id: usuarioId
     });
     return distributorId;
