+++ src/controler/controlerProduto.js
@@ -1,0 +1,3 @@
+import { kknex } from "../config/database.js";
+
+ //vai adicionar os produtos
