--- src/types.d.ts
@@ -1,7 +1,0 @@
-import { Request } from 'express';
-
-declare module 'express-serve-static-core' {
-  interface Request {
-    userId?: number;
-  }
-}
